🔨 refactor: my place feature 분리
