+++ src/features/my-place/my-place-common/components/my-place-item.tsx
